u-boot: armv7: fix dcache disable issue on G1E with new toolchain(gcc 7.2)
The problem is there is one cache-cable memory access between L1 and L2 cache flush.
and the process of disabling cache and MMU did not complete yet.

Due to above issue, the memory access is not exist at this time, so system
can not jump to linux(kernel). And the address of device tree is NULL.
And it leads to issue 'Unsupported Architecture 0x0'.

Currently, the issue 'Unsupported Architecture 0x0' only occurs on G1E
with new toolchain(gcc 7.2).

This patch refer commit:
https://git.rigado.com/vesta/u-boot-2015.04/commit/d85cd484e6825631aa1ab572e5e0539f2191d795?view=inline

Signed-off-by: khatran <kha.tran.px@rvc.renesas.com>