board: silinux: ek874: add function to check board revision
This commit is created to add a new function to check board revision
and use right devicetree files corresponding with board revisions.
Currently, we can monitor the value of GP5_19 pin to identify
EK874 between Rev.C and the latest Rev.E.

In latest EK874 Revisions, this pin is pulled down with an external
4.7kOhm resistor. Therefore, when we enable internal pull-up then
setting general input mode for this pin, we will find the different
value of GP5_19 in each revisions:
- Rev.C and the older: 1
- The latest Rev.E: 0

And Rev.B has the SoC rev1.0 while others are rev1.1.

Signed-off-by: hienhuynh <hien.huynh.px@renesas.com>