mmc: sdhi: flexible support setting for mutual channel SDHI-IF
Some SDHI-IF channels can support both eMMC and SD interfaces by
dip-switch selection in EVK Board such as RZ/G2L Smarc Series...
But currently, it can not switch flexibly and hard set interface in
its devicetree.
If we want to use eMMC or SD interfaces, we have to set the macro
definition to choose mode in devicetree then rebuild it.
This is not flexible and cost us time to do this.
So to support flexible setting, we create a devicetree property called
"mutual-channel" to notify driver that this channel can support eMMC
and SD interfaces.

Signed-off-by: tranhha <tranh.ha.xb@renesas.com>