usb: gadget: rcar: add support for RZ/G2E
This patch adds support for RZ/G2E because it is as same as RCAR D3.

This SoC needs to release the PLL reset by the UGCTRL register.
So, since this is not the same as other R-Car Gen3 SoCs,
this patch adds a new type as "USBHS_TYPE_RCAR_GEN3_WITH_PLL".

Signed-off-by: Huy Nguyen <huy.nguyen.wh@renesas.com>