configs: rzf: let mmc user partition store env instead of boot partition
SDHI0-IF can support both eMMC and uSD interfaces in RZ/Five Dev and
Smarc Boards.
Currently, if we switch to use uSD instead of eMMC, u-boot environment
can not be saved becaused saved partition is boot partition which uSD
card does not have this partition.
So, move to use user partition to store u-boot environment to let it
work normally in both eMMC and uSD.

Signed-off-by: Hien Huynh <hien.huynh.px@renesas.com>