mmc: sh_sdhi: fix fail to boot sd card for RZ/N2H
After rebooting the board, an error occurred:
 ** fs_devread read error - block
Failed to load 'boot/Image'
** No partition table - mmc 1 **
Couldn't find partition mmc 1:2
Can't set block device

The reason for this error is that after rebooting the board,
SD1_PWEN is not active, causing SD1 to be undetectable.

There's commit fixed this issue on RZ/T2H. But on RZ/N2H,
we need to increase mdelay to 6 to avoid this error.

Signed-off-by: Nhat Nguyen <nhat.nguyen.xb@renesas.com>